fixed the monitor unit testing
